requirements.txt
Django>=5.0.2,<5.1
djangorestframework>=3.15.1,<3.16
psycopg[c]>=3.1.18,<3.2
drf-spectacular==0.27.2
#Pillow>=9.1.0,<9.2
#uwsgi>=2.0.24,<2.1
django-localflavor>=4.0,<5.0
